+++ R/print.sc.R
@@ -8,21 +8,27 @@
 NULL
 
 #' @rdname print.sc
-#' @param duration If TRUE the duration for computation is printed.
 #' @export
-print.sc_power <- function(x, duration = FALSE, ...) {
+print.sc_power <- function(x, digits = "auto", ...) {
 
-  cat("Test-Power in percent:\n\n")
-  
-  class(x) <- "data.frame"
-  print(x,row.names = FALSE)
-  if (duration) 
-    cat(
-      "\nComputation duration is", 
-      round(attr(x, "computation_duration")[3], 1), 
-      "seconds.\n"
+  cat("Test-Power in percent:\n")
+  ma <- matrix(
+    unlist(x[1:16]) * 100, byrow = FALSE, ncol = 2, 
+    dimnames = list(
+      c(
+        "tauU: A vs. B - Trend A", 
+        paste0("Rand-Test: ",x$rand.test.stat[1]),  
+        "PLM.Norm: Level", 
+        "PLM.Norm: Slope", 
+        "PLM.Poisson: Level", 
+        "PLM.Poisson: Slope", 
+        "HPLM: Level", 
+        "HPLM: Slope"
+      ), 
+      c("Power", "Alpha-error")
+    )
   )
-  
+  ma
 }
 
 
@@ -81,19 +87,15 @@ print.sc_cdc <- function(x, nice = TRUE, ...) {
 print.sc_bctau <- function(x, nice = TRUE, digits = "auto", ...) {
   
   cat("Baseline corrected tau\n\n")
+  cat("\n")
   
-  if (x$repeated) {
-    cat("Method: Siegel repeated median regression\n")
-  } else {
-    cat("Method: Theil-Sen regression\n")
-  }
   
   if (x$continuity) {
     cat("Continuity correction applied\n")
   } else {
     cat("Continuity correction not applied.\n")
   }
-  cat("\n")
+  
   if (digits == "auto") {
     x$parameters$p <- round(x$parameters$p, 3)
     x$parameters$z <- sprintf("%.2f", x$parameters$z)
@@ -144,6 +146,7 @@ print.sc_desc <- function(x, digits = "auto", ...) {
 print.sc_design <- function(x, ...) {
   cat("A scdf design matrix\n\n")
   cat("Number of cases:", length(x$cases), "\n")
+<<<<<<< HEAD
   cat("Distribution: ", x$distribution, "\n")
   cat("Start values: ", unique(
     sapply(x$cases, function(x) {x$start_value[1]})), "\n")
@@ -171,6 +174,21 @@ print.sc_design <- function(x, ...) {
     })), "\n")
   }
   
+=======
+  cat("Mean: ", x$cases[[1]]$m[1], "\n")
+  cat("SD = ", x$cases[[1]]$s[1], "\n")
+  cat("rtt = ", x$cases[[1]]$rtt[1], "\n")
+  cat("Phase design: ", as.character(x$cases[[1]]$phase), "\n")
+  
+  cat("mean trend-effect: ", apply(sapply(x$cases, function(x) {x$trend}), 1, mean, na.rm = TRUE)[1], "\n")
+  cat("mean level-effect: ", apply(sapply(x$cases, function(x) {x$level}), 1, mean, na.rm = TRUE), "\n")
+  cat("mean slope-effect: ", apply(sapply(x$cases, function(x) {x$slope}), 1, mean, na.rm = TRUE), "\n")
+  cat("sd trend-effect: ", apply(sapply(x$cases, function(x) {x$trend}), 1, sd, na.rm = TRUE)[1], "\n")
+  cat("sd level-effect: ", apply(sapply(x$cases, function(x) {x$level}), 1, sd, na.rm = TRUE), "\n")
+  cat("sd slope-effect: ", apply(sapply(x$cases, function(x) {x$slope}), 1, sd, na.rm = TRUE), "\n")
+  cat("Distribution: ", x$distribution)
+
+>>>>>>> master
 }
 
 #' @rdname print.sc
@@ -466,13 +484,10 @@ print.sc_tauu <- function(x, complete = FALSE, digits = "auto", ...) {
   cat("Tau-U\n")
   cat("Method:", x$method, "\n")
   cat("Applied Kendall's Tau-", x$tau_method, "\n", sep = "")
-  if (complete || (length(x$table) > 1 && x$meta_method != "none")) {
-    cat(x$ci * 100, "% CIs for tau are reported.\n\n", sep = "")
-  } else cat("\n")
-  
+  cat(x$ci * 100, "% CIs for tau are reported.\n\n", sep = "")
   out <- x$table
   
-  if (length(out) > 1 && x$meta_method != "none") {
+  if (length(out) > 1) {
     cat("Overall Tau-U\n")
     cat("Meta-anlysis model:", x$meta_method, "effect\n\n")
     print(x$Overall_tau_u, row.names = FALSE, digits = digits)
@@ -514,11 +529,11 @@ print.sc_plm <- function(x, ...) {
     cat("Correlated residuals up to autoregressions of lag",
         x$ar, "are modelled\n\n")
   
-  if (x$family == "poisson" || x$family == "binomial") {
+  if (x$family == "poisson" || x$family == "nbinomial") {
     Chi <- x$full$null.deviance - x$full$deviance
     DF <- x$full$df.null - x$full$df.residual
     cat(sprintf(
-      "X\u00b2(%d) = %.2f; p = %0.3f; AIC = %.0f\n\n", 
+      "\u0347\u00b2(%d) = %.2f; p = %0.3f; AIC = %.0f\n\n", 
       DF, Chi, 1 - pchisq(Chi, df = DF), x$full$aic)
     )	
   } else {
@@ -561,7 +576,7 @@ print.sc_plm <- function(x, ...) {
   if (is.null(x$r.squares))
     colnames(res) <- c("B", "2.5%", "97.5%", "SE", "t", "p")		
   
-  if (x$family == "poisson" || x$family == "binomial") {
+  if (x$family == "poisson" || x$family == "nbinomial") {
     OR <- exp(res[, 1:3])
     Q <- (OR - 1) / (OR + 1)
     res <- cbind(res[, -7], round(OR, 3), round(Q, 2))
@@ -572,22 +587,14 @@ print.sc_plm <- function(x, ...) {
   }
   print(res)
   cat("\n")
-  if (x$family == "gaussian") {
-    cat("Autocorrelations of the residuals\n")
-    lag.max = 3
-    cr <- acf(residuals(x$full.model), lag.max = lag.max,plot = FALSE)$acf[2:(1 + lag.max)]
-    cr <- round(cr, 2)
-    print(data.frame(lag = 1:lag.max, cr = cr), row.names = FALSE)
-    cat("\n")
-  }
+  cat("Autocorrelations of the residuals\n")
+  lag.max = 3
+  cr <- acf(residuals(x$full.model), lag.max = lag.max,plot = FALSE)$acf[2:(1 + lag.max)]
+  cr <- round(cr, 2)
+  print(data.frame(lag = 1:lag.max, cr = cr), row.names = FALSE)
+  cat("\n")
   cat("Formula: ")
-  if (x$family == "binomial") {
-    x$formula[2] <- str2expression(paste0(x$formula[2], "/", x$var_trials))
-  }
   print(x$formula, showEnv = FALSE)
-  if (x$family == "binomial") {
-    cat("weights = ", x$var_trials)
-  }
   cat("\n")
   .note_vars(x)
 }
@@ -690,6 +697,19 @@ print.sc_rci <- function(x, ...) {
   cat("\n")
 }
 
+
+.note_vars <- function(x) {
+  v <- any(attr(x, .opt$dv) != "values")
+  p <- attr(x, .opt$phase) != "phase"
+  m <- attr(x, .opt$mt) != "mt"
+  if (v || p || m) { 
+    cat("\nThe following variables were used in this analysis:\n'", 
+        paste0(attr(x, .opt$dv), collapse = "/ "), "' as dependent variable, '", 
+        paste0(attr(x, .opt$phase), collapse = "/ "), "' as phase variable, and '", 
+        paste0(attr(x, .opt$mt), collapse = "/ "),"' as measurement-time variable.\n", sep = "")
+  }
+}
+
 #' @rdname print.sc
 #' @export
 print.sc_smd <- function(x, digits = "auto", ...) {
@@ -706,17 +726,3 @@ print.sc_smd <- function(x, digits = "auto", ...) {
   .note_vars(x)
   
 }
-
-
-.note_vars <- function(x) {
-  v <- any(attr(x, .opt$dv) != "values")
-  p <- attr(x, .opt$phase) != "phase"
-  m <- attr(x, .opt$mt) != "mt"
-  if (v || p || m) { 
-    cat("\nThe following variables were used in this analysis:\n'", 
-        paste0(attr(x, .opt$dv), collapse = "/ "), "' as dependent variable, '", 
-        paste0(attr(x, .opt$phase), collapse = "/ "), "' as phase variable, and '", 
-        paste0(attr(x, .opt$mt), collapse = "/ "),"' as measurement-time variable.\n", sep = "")
-  }
-}
-