+++ R/pem.R
@@ -22,7 +22,7 @@
 #' @examples
 #' 
 #' ## Calculate the PEM including the Binomial and Chi-square tests for a single-case
-#' dat <- rSC(5, level = 0.5)
+#' dat <- random_scdf(5, level = 0.5)
 #' pem(dat, chi.test = TRUE)
 #' 
 #' @export