+++ R/plot.scdf.R
@@ -87,7 +87,7 @@
 #'        lines = list(type = "loreg", f = 0.2, lty = "solid", col = "black", lwd = 3))
 #' 
 #' ## Plot a random MBD over three cases and mark interesting MTs
-#' dat <- rSC(design = design_rSC(3))
+#' dat <- random_scdf(design = design(3))
 #' plot(dat, marks = list(positions = list(c(2,4,5),c(1,2,3),c(7,8,9)), col = "blue",
 #'        cex = 1.4), style = c("grid", "annotate", "tiny"))
 #' 
@@ -129,11 +129,11 @@ plotSC <- function(data, dvar, pvar, mvar,
   if (is.list(style)) {
     ref.style <- "default"
     if ("style" %in% names(style)) ref.style <- style$style
-    style <- c(style, style_plotSC(ref.style))
+    style <- c(style, style_plot(ref.style))
     style <- style[unique(names(style))]
   }
   
-  if (is.character(style)) style <- style_plotSC(style)
+  if (is.character(style)) style <- style_plot(style)
   
   #for pre style backwards compatibility
   sty_names <- c("fill", "fill.bg", "frame", "grid", "lwd", "pch", "text.ABlag", "type")
@@ -163,7 +163,7 @@ plotSC <- function(data, dvar, pvar, mvar,
   
   ### END: define style
   
-  # Marks on the outliers from outlierSC
+  # Marks on the outliers from outlierandom_scdf
   if (identical(class(marks), c("sc_outlier"))) 
     marks <- list(positions = marks$dropped.mt)