+++ R/private_prepare_scdf.R
@@ -1,11 +1,5 @@
 
-.prepare_scdf <- function(data, 
-                           na.rm = FALSE 
-                           #change.var.phase = FALSE, 
-                           #change.var.values = FALSE, 
-                           #change.var.mt = FALSE
-                           ) {
-  
+.prepare_scdf <- function(data, na.rm = FALSE) {
   
   # check class scdf validity
   if (.opt$rigorous_class_check) {
@@ -38,8 +32,6 @@
     if (!is.factor(data[[case]][[pvar]])) {
       data[[case]][[pvar]] <- as.factor(data[[case]][[pvar]])
     }
-    
-    #if (is.na(names(data)[case])) names(data)[case] <- paste0("Case ", case)
   }
   data
 }