+++ R/scan.R
@@ -12,11 +12,11 @@
     scan.print.cases = "fit",
     scan.print.rows   = 15,
     scan.print.cols   = "all",
-    scan.print.digits = 1,
+    scan.print.digits = 2,
     scan.print.long   = FALSE,
     scan.print.scdf.name = TRUE,
     scan.plot.style = "grid",
-    scan.deprecated.warning = FALSE,
+    scan.deprecated.warning = TRUE,
     scan.export.kable = list(digits = 2, linesep ="", booktab = TRUE),
     scan.export.kable_styling = list(
       bootstrap_options = c("bordered", "condensed"),