+++ R/smd.R
@@ -3,6 +3,15 @@
 #' The \code{smd} function provides various standardized mean effect sizes for single-case data.
 #'
 #' @inheritParams .inheritParams
+#' @details  
+#' 'sd cohen' is the (unweigted) average of the variance of phase A and B.  
+#' 'sd Hedges' is the weighted average of the variance of phase A and B (with a degrees of freedom correction). 
+#' 'Hedges' g' is the mean difference divided by 'sd Hedges'.
+#' 'Hedges' g correction' and 'Hedges' g durlak correction' are two approaches 
+#' of correcting Hedges' g for small sample sizes.
+#' 'Glass' delta' is the mean difference divided by the standard deviation of the 
+#' A-phase.
+#' 'Cohens d` is the mean difference divided by 'sd cohen'.
 #' @author Juergen Wilbert
 #' @seealso \code{\link{overlap}}, \code{\link{describe}}
 #' @export