+++ R/style_plot.R
@@ -86,14 +86,14 @@ style_plot <- function(style = "default", ...) {
   invisible(out)
 }
 
-#' @rdname style_plot
+#' @rdname deprecated-functions
 #' @export
 style_plotSC <- function(...) {
   .deprecated_warning("style_plot", "style_plotSC")
   style_plot(...)
 }
 
-#' @rdname style_plot
+#' @rdname deprecated-functions
 #' @export
 style.plotSC <- function(...) {
   .deprecated_warning("style_plot", "style.plotSC")